restructure write barrier as part of experimental GC
